refactor: update SVG import paths to use consistent naming convention and remove deprecated icons
