+++ src/common/Navigation/Search/styled.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { ReactComponent as IconSearch } from "../../../images/iconSearch.svg";
+import { ReactComponent as IconSearch } from "../../../images/Search.svg";
 
 export const SearchContainer = styled.div`
   display: flex;
